+++ src/main/java/roomescape/member/controller/MemberController.java
@@ -1,13 +1,20 @@
 package roomescape.member.controller;
 
-import org.springframework.stereotype.Controller;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.member.auth.dto.MemberInfo;
+import roomescape.member.service.MemberService;
 
-@Controller
+@RequiredArgsConstructor
+@RestController
 public class MemberController {
 
-    @GetMapping("/reservation")
-    public String getReservationPage() {
-        return "reservation";
+    private final MemberService memberService;
+
+    @GetMapping("/members")
+    public List<MemberInfo> getMembers() {
+        return memberService.getAll();
     }
 }
