--- src/main/java/roomescape/member/controller/UserController.java
@@ -1,32 +1,0 @@
-package roomescape.member.controller;
-
-import jakarta.servlet.http.HttpServletRequest;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
-import roomescape.member.controller.dto.LoginCheckResponse;
-import roomescape.member.controller.dto.LoginRequest;
-import roomescape.member.controller.dto.SignupRequest;
-import roomescape.member.domain.Member;
-import roomescape.member.service.MemberService;
-import roomescape.member.service.AuthService;
-
-@RequiredArgsConstructor
-@Controller
-public class UserController {
-
-    @GetMapping("/login")
-    public String getLoginPage() {
-        return "login";
-    }
-
-    @GetMapping("/signup")
-    public String getSignupPage() {
-        return "signup";
-    }
-}
