+++ src/main/java/roomescape/member/controller/ui/UserController.java
@@ -1,0 +1,26 @@
+package roomescape.member.controller.ui;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@RequiredArgsConstructor
+@Controller
+public class UserController {
+
+    @GetMapping("/login")
+    public String getLoginPage() {
+        return "login";
+    }
+
+    @GetMapping("/signup")
+    public String getSignupPage() {
+        return "signup";
+    }
+
+    // require authentication
+    @GetMapping("/reservation")
+    public String getReservationPage() {
+        return "reservation";
+    }
+}
