+++ src/main/java/roomescape/member/repository/MemberRepository.java
@@ -1,5 +1,6 @@
 package roomescape.member.repository;
 
+import java.util.List;
 import java.util.Optional;
 import roomescape.member.domain.Account;
 import roomescape.member.domain.Member;
@@ -14,4 +15,6 @@ public interface MemberRepository {
     Optional<Member> findById(MemberId id);
 
     Optional<Account> findAccountByEmail(MemberEmail email);
+
+    List<Member> findAll();
 }
