+++ src/main/java/roomescape/reservation/controller/dto/ReservationWebResponse.java
@@ -3,6 +3,7 @@ package roomescape.reservation.controller.dto;
 import lombok.AccessLevel;
 import lombok.experimental.FieldNameConstants;
 import roomescape.common.utils.Validator;
+import roomescape.member.auth.dto.MemberInfo;
 import roomescape.theme.controller.dto.ThemeWebResponse;
 import roomescape.time.controller.dto.ReservationTimeWebResponse;
 
@@ -10,23 +11,23 @@ import java.time.LocalDate;
 
 @FieldNameConstants(level = AccessLevel.PRIVATE)
 public record ReservationWebResponse(Long id,
-                                     String name,
+                                     MemberInfo memberInfo,
                                      LocalDate date,
                                      ReservationTimeWebResponse time,
                                      ThemeWebResponse theme) {
 
     public ReservationWebResponse {
-        validate(id, name, date, time, theme);
+        validate(id, memberInfo, date, time, theme);
     }
 
     private void validate(final Long id,
-                          final String name,
+                          final MemberInfo memberInfo,
                           final LocalDate date,
                           final ReservationTimeWebResponse time,
                           final ThemeWebResponse theme) {
         Validator.of(ReservationWebResponse.class)
                 .notNullField(Fields.id, id)
-                .notBlankField(Fields.name, name)
+                .notNullField(Fields.memberInfo, memberInfo)
                 .notNullField(Fields.date, date)
                 .notNullField(Fields.time, time)
                 .notNullField(Fields.theme, theme);
